refactor: extract vite plugins to separate files
