+++ packages/@storybook/astro/src/viteStorybookAstroMiddlewarePlugin.ts
@@ -3,6 +3,7 @@ import { createServer, type PluginOption } from 'vite';
 import type { RenderRequestMessage, RenderResponseMessage } from '@storybook/astro-renderer/types';
 import type { FrameworkOptions } from './types';
 import type { Integration } from './integrations';
+import { viteAstroContainerRenderersPlugin } from './viteAstroContainerRenderersPlugin';
 
 export async function vitePluginStorybookAstroMiddleware(options: FrameworkOptions) {
   const viteServer = await createViteServer(options.integrations);
@@ -69,92 +70,9 @@ export async function createViteServer(integrations: Integration[]) {
     ...config,
     plugins: [
       ...(config.plugins?.filter(Boolean) ?? []),
-      astroRenderersPlugin(integrations)
+      viteAstroContainerRenderersPlugin(integrations)
     ]
   });
 
   return viteServer;
-}
-
-function astroRenderersPlugin(integrations: Integration[]) {
-  const name = 'astro-renderers';
-  const virtualModuleId = `virtual:${name}`;
-  const resolvedVirtualModuleId = `\0${virtualModuleId}`;
-
-  return {
-    name,
-
-    resolveId(id: string) {
-      if (id === virtualModuleId) {
-        return resolvedVirtualModuleId;
-      }
-    },
-
-    load(id: string) {
-      if (id === resolvedVirtualModuleId) {
-        const importStatements = buildImportStatements(integrations);
-
-        return `
-          ${importStatements}
-          export function addRenderers(container) {
-            ${integrations.map(integration => buildServerRenderer(integration)).join('\n')}
-            ${integrations.map(integration => buildClientRenderer(integration)).join('\n')}
-          }
-        `;
-      }
-    }
-  };
-}
-
-function buildImportStatements(integrations: Integration[]) {
-  return integrations
-    .filter((integration) => integration.renderer.server)
-    .map(
-      (integration) =>
-        `import ${integration.name}Renderer from '${integration.renderer.server?.entrypoint}';`
-    )
-    .join('\n');
-}
-
-function buildServerRenderer(integration: Integration) {
-  const serverRenderer = integration.renderer.server;
-
-  if (!serverRenderer) {
-    return '';
-  }
-
-  if (integration.name === 'solid') {
-    return `
-      container.addServerRenderer({
-        name: '${serverRenderer.name}',
-        renderer: {
-          ...${integration.name}Renderer,
-          name: '${serverRenderer.name}'
-        }
-      });
-    `;
-  }
-
-  
-  return `
-    container.addServerRenderer({
-      name: '${serverRenderer.name}',
-      renderer: ${integration.name}Renderer
-    });
-  `;
-}
-
-function buildClientRenderer(integration: Integration) {
-  const clientRenderer = integration.renderer.client;
-  
-  if (clientRenderer) {
-    return `
-      container.addClientRenderer({
-        name: '${clientRenderer.name}',
-        renderer: '${clientRenderer.entrypoint}'
-      });
-      `;
-  }
-
-  return '';
 }
